+++ sereto/cli/cli.py
@@ -4,7 +4,6 @@ from pathlib import Path
 import click
 import keyring
 
-from sereto.cleanup import render_sow_cleanup
 from sereto.cli.commands import sereto_ls, sereto_repl
 from sereto.cli.utils import AliasedGroup, Console
 from sereto.config import (
@@ -651,8 +650,7 @@ def pdf_sow(
     Console().log(f"Rendering SoW version: '{version}'")
     report_create_missing(report=report, settings=settings, version=version)
     render_sow_j2(report=report, settings=settings, version=version)
-    render_sow_pdf(settings=settings, version=version, recipe=sow_recipe)
-    render_sow_cleanup(settings=settings, version=version)
+    render_sow_pdf(settings=settings, version=version, recipe=sow_recipe, keep_original=False)
 
 
 # -------------