+++ sereto/types.py
@@ -47,7 +47,7 @@ Example:
 """
 
 
-TypePassword = Annotated[str, Strict, Len(8, 100)]
+TypePassword = Annotated[str, Strict(), Len(8, 100)]
 """Type for password.
 
 The value should meet the following requirements:
@@ -62,14 +62,14 @@ def zero_bytes(value: bytes) -> bytes:
     return value
 
 
-TypeNonce12B = Annotated[bytes, Len(12, 12), Strict, AfterValidator(zero_bytes)]
+TypeNonce12B = Annotated[bytes, Len(12, 12), Strict(), AfterValidator(zero_bytes)]
 """Type for a 12 byte long nonce.
 
 The value must contain at least one non-zero byte. This check is in place to prevent unintentional errors.
 """
 
 
-TypeSalt16B = Annotated[bytes, Len(16, 16), Strict, AfterValidator(zero_bytes)]
+TypeSalt16B = Annotated[bytes, Len(16, 16), Strict(), AfterValidator(zero_bytes)]
 """Type for a 16 byte long salt.
 
 The value must contain at least one non-zero byte. This check is in place to prevent unintentional errors.